Fix anti-cheat build errors and incorrect XOR blacklist values
- Fix all XOR-encrypted process/window blacklist values (were computing
  to garbage, never matching any cheat tool name)
- Add r3dAntiCheat files to Studio.vcproj to fix LNK2019 unresolved externals
- Fix MSVC 2008 cast errors: use (Playerstate_e)(int) for r3dSec_type
- Fix MSVC 2008 compound assignment: convert -= and += to explicit form
  since older compiler cannot resolve template operators

+++ src/EclipseStudio/Sources/UI/HUD_TPSGame.cpp
@@ -1438,7 +1438,7 @@ bool g_AutoRunToggled = false;
 
 void ProcessCamera(obj_Player* pl)
 {
-	ActiveCameraRigID = (Playerstate_e)pl->PlayerState;
+	ActiveCameraRigID = (Playerstate_e)(int)pl->PlayerState;
 	ActiveCameraRig   = TPSHudCameras[g_camera_mode->GetInt()][ActiveCameraRigID];
 	
 	// use this to update the camera from the options. 
@@ -1468,7 +1468,7 @@ void ProcessCamera(obj_Player* pl)
 
 		LerpValue = 0;
 
-		CurrentState = (Playerstate_e)pl->PlayerState;	
+		CurrentState = (Playerstate_e)(int)pl->PlayerState;
 	}
 	else
 	{
